move auspice export code to seperate file
